+++ src/datagen/java/blusunrize/immersiveengineering/data/tags/IEBlockTags.java
@@ -277,8 +277,8 @@ public class IEBlockTags extends BlockTagsProvider
 			tag(BlockTags.SLABS).add(slab.get());
 		for(BlockEntry<?> stairs : IEBlocks.TO_STAIRS.values())
 			tag(BlockTags.STAIRS).add(stairs.get());
-		for(BlockEntry<?> stairs : IEBlocks.TO_WALL.values())
-			tag(BlockTags.WALLS).add(stairs.get());
+		for(BlockEntry<?> walls : IEBlocks.TO_WALL.values())
+			tag(BlockTags.WALLS).add(walls.get());
 
 		tag(IETags.incorrectDropsSteel).addTag(BlockTags.INCORRECT_FOR_IRON_TOOL);